feat(batch-tx): send_fee_selection 화면에서 수수료 조회

+++ lib/providers/view_model/send/send_address_view_model.dart
@@ -91,7 +91,7 @@ class SendAddressViewModel extends ChangeNotifier {
         totalSendAmount;
   }
 
-  void saveWalletIdAndBatchRecipients(int id, Map<String, int> recipients) {
+  void saveWalletIdAndBatchRecipients(int id, Map<String, double> recipients) {
     _sendInfoProvider.setWalletId(id);
     _sendInfoProvider.setRecipientsForBatch(recipients);
   }
